refactor: rename Features/NoneResponse to Features/NoneFlow
Mirrors the sibling Features/ImplicitFlow naming. Folder and namespace
only; NoneResponseBuilder keeps its name, matching the response-type
naming of IdTokenResponseBuilder/TokenResponseBuilder in ImplicitFlow.
